# Added benchmark and optimised ParseBytes