The first optimisation is to pre-emptively grow the buffer to the
right size. The second is to more efficiently convert a byte to a zero
padded 8 bit binary string. byteToBinary returns a `[8]byte` instead of a
`[]byte` to prevent it escaping to the heap.

```
benchmark              old ns/op      new ns/op      delta
BenchmarkParse/0-4     5996           2482           -58.61%
BenchmarkParse/1-4     10321          2891           -71.99%
BenchmarkParse/2-4     9658           2963           -69.32%
BenchmarkParse/3-4     10338          3082           -70.19%
BenchmarkParse/4-4     11903          3318           -72.12%
BenchmarkParse/5-4     12563          3176           -74.72%

benchmark              old allocs     new allocs     delta
BenchmarkParse/0-4     40             16             -60.00%
BenchmarkParse/1-4     64             17             -73.44%
BenchmarkParse/2-4     66             17             -74.24%
BenchmarkParse/3-4     70             19             -72.86%
BenchmarkParse/4-4     74             19             -74.32%
BenchmarkParse/5-4     72             19             -73.61%

benchmark              old bytes      new bytes      delta
BenchmarkParse/0-4     1592           1164           -26.88%
BenchmarkParse/1-4     2496           1508           -39.58%
BenchmarkParse/2-4     2512           1508           -39.97%
BenchmarkParse/3-4     2624           1636           -37.65%
BenchmarkParse/4-4     2688           1700           -36.76%
BenchmarkParse/5-4     2640           1636           -38.03%
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/liveramp/iabconsent/3)
<!-- Reviewable:end -->