@nylar 2018-05-25
Master is currently outperforming this branch, so I am going to close it:

```shell
BenchmarkParse/0-4 	 1000000	      1535 ns/op	     588 B/op	      12 allocs/op
BenchmarkParse/1-4 	 1000000	      1454 ns/op	     564 B/op	      13 allocs/op
BenchmarkParse/2-4 	 1000000	      1480 ns/op	     564 B/op	      13 allocs/op
BenchmarkParse/3-4 	 1000000	      1617 ns/op	     636 B/op	      14 allocs/op
BenchmarkParse/4-4 	 1000000	      1590 ns/op	     636 B/op	      14 allocs/op
BenchmarkParse/5-4 	 1000000	      1560 ns/op	     636 B/op	      14 allocs/op
```